Module metaflac::block

Includes various types of metadata blocks.

Structs

Application
A structure representing an APPLICATION block.

CueSheet
A structure representing a CUESHEET block.

CueSheetTrack
A structure representing a cuesheet track.

CueSheetTrackIndex
A structure representing a cuesheet track index.

Picture
A structure representing a PICTURE block.

SeekPoint
A structure representing a seektable seek point.

SeekTable
A structure representing a SEEKTABLE block.

StreamInfo
A structure representing a STREAMINFO block.

VorbisComment
A structure representing a VORBIS_COMMENT block.

Enums

Block
The parsed content of a metadata block.

BlockType
Types of blocks. Used primarily to map blocks to block identifiers when reading and writing.

PictureType
Types of pictures that can be used in the picture block.