Enum metaflac::block::Block

pub enum Block {
    StreamInfo(StreamInfo),
    Application(Application),
    CueSheet(CueSheet),
    Padding(u32),
    Picture(Picture),
    SeekTable(SeekTable),
    VorbisComment(VorbisComment),
    Unknown((u8, Vec<u8>)),
}

The parsed content of a metadata block.

Variants

StreamInfo

A value containing a parsed streaminfo block.

Application

A value containing a parsed application block.

CueSheet

A value containing a parsed cuesheet block.

Padding

A value containing the number of bytes of padding.

Picture

A value containing a parsed picture block.

SeekTable

A value containing a parsed seektable block.

VorbisComment

A value containing a parsed vorbis comment block.

Unknown

An value containing the bytes of an unknown block.

Methods

impl Block

fn read_from(reader: &mut Read) -> Result<(bool, u32, Block)>

Attempts to read a block from the reader. Returns a tuple containing a boolean indicating if the block was the last block, the length of the block in bytes, and the new Block.

fn write_to(&self, is_last: bool, writer: &mut Write) -> Result<u32>

Attemps to write the block to the writer. Returns the length of the block in bytes.

fn block_type(&self) -> BlockType

Returns the corresponding block type byte for the block.