Struct metaflac::block::CueSheetTrack

pub struct CueSheetTrack {
    pub offset: u64,
    pub number: u8,
    pub isrc: String,
    pub is_audio: bool,
    pub pre_emphasis: bool,
    pub indices: Vec<CueSheetTrackIndex>,
}

A structure representing a cuesheet track.

Fields

offset

Track offset in samples, relative to the beginning of the FLAC audio stream. It is the offset to the first index point of the track.

number

Track number.

isrc

Track ISRC. This is a 12-digit alphanumeric code.

is_audio

The track type.

pre_emphasis

The pre-emphasis flag.

indices

For all tracks except the lead-out track, one or more track index points.

Methods

impl CueSheetTrack

fn new() -> CueSheetTrack

Returns a new CueSheetTrack of type audio, without pre-emphasis, and with zero/empty values.