Struct metaflac::block::Picture

pub struct Picture {
    pub picture_type: PictureType,
    pub mime_type: String,
    pub description: String,
    pub width: u32,
    pub height: u32,
    pub depth: u32,
    pub num_colors: u32,
    pub data: Vec<u8>,
}

A structure representing a PICTURE block.

Fields

picture_type

The picture type.

mime_type

The MIME type.

description

The description of the picture.

width

The width of the picture in pixels.

height

The height of the picture in pixels.

depth

The color depth of the picture in bits-per-pixel.

num_colors

For indexed-color pictures (e.g. GIF), the number of colors used, or 0 for non-indexed pictures.

data

The binary picture data.

Methods

impl Picture

fn new() -> Picture

Returns a new Picture with zero/empty values.

fn from_bytes(bytes: &[u8]) -> Result<Picture>

Attempts to parse the bytes as a Picture block. Returns a Picture on success.

fn to_bytes(&self) -> Vec<u8>

Returns a vector representation of the picture block suitable for writing to a file.