Struct metaflac::block::VorbisComment

pub struct VorbisComment {
    pub vendor_string: String,
    pub comments: HashMap<String, Vec<String>>,
}

A structure representing a VORBIS_COMMENT block.

Fields

vendor_string

The vendor string.

comments

A map of keys to a list of their values.

Methods

impl VorbisComment

fn new() -> VorbisComment

Returns a new VorbisComment with an empty vendor string and no comments.

fn from_bytes(bytes: &[u8]) -> Result<VorbisComment>

Attempts to parse the bytes as a vorbis comment block. Returns a VorbisComment on success.

fn to_bytes(&self) -> Vec<u8>

Returns a vector representation of the vorbis comment suitable for writing to a file.

fn get(&self, key: &str) -> Option<&Vec<String>>

Returns a reference to the vector of comments for the specified key.

fn set<K: Into<String>, V: Into<String>>(&mut self, key: K, values: Vec<V>)

Sets the comments for the specified key. Any previous values under the key will be removed.

fn remove(&mut self, key: &str)

Removes the comments for the specified key.

fn remove_pair(&mut self, key: &str, value: &str)

Removes any matching key/value pairs.

fn artist(&self) -> Option<&Vec<String>>

Returns a reference to the vector of values with the ARTIST key.

fn set_artist<T: Into<String>>(&mut self, artists: Vec<T>)

Sets the values for the ARTIST key. This will result in any ARTISTSORT comment being removed.

fn remove_artist(&mut self)

Removes all values with the ARTIST key. This will result in any ARTISTSORT comments being removed as well.

fn album(&self) -> Option<&Vec<String>>

Returns a reference to the vector of values with the ALBUM key.

fn set_album<T: Into<String>>(&mut self, albums: Vec<T>)

Sets the values for the ALBUM key. This will result in any ALBUMSORT comments being removed.

fn remove_album(&mut self)

Removes all values with the ALBUM key. This will result in any ALBUMSORT comments being removed as well.

fn genre(&self) -> Option<&Vec<String>>

Returns a reference to the vector of values with the GENRE key.

fn set_genre<T: Into<String>>(&mut self, genres: Vec<T>)

Sets the values for the GENRE key.

fn remove_genre(&mut self)

Removes all values with the GENRE key.

fn title(&self) -> Option<&Vec<String>>

Returns reference to the vector of values with the TITLE key.

fn set_title<T: Into<String>>(&mut self, title: Vec<T>)

Sets the values for the TITLE key. This will result in any TITLESORT comments being removed.

fn remove_title(&mut self)

Removes all values with the TITLE key. This will result in any TITLESORT comments being removed as well.

fn track(&self) -> Option<u32>

Attempts to convert the first TRACKNUMBER comment to a u32.

fn set_track(&mut self, track: u32)

Sets the TRACKNUMBER comment.

fn remove_track(&mut self)

Removes all values with the TRACKNUMBER key.

fn total_tracks(&self) -> Option<u32>

Attempts to convert the first TOTALTRACKS comment to a u32.

fn set_total_tracks(&mut self, total_tracks: u32)

Sets the TOTALTRACKS comment.

fn remove_total_tracks(&mut self)

Removes all values with the TOTALTRACKS key.

fn album_artist(&self) -> Option<&Vec<String>>

Returns a reference to the vector of values with the ALBUMARTIST key.

fn set_album_artist<T: Into<String>>(&mut self, album_artists: Vec<T>)

Sets the values for the ALBUMARTIST key. This will result in any ALBUMARTISTSORT comments being removed.

fn remove_album_artist(&mut self)

Removes all values with the ALBUMARTIST key. This will result in any ALBUMARTISTSORT comments being removed as well.

fn lyrics(&self) -> Option<&Vec<String>>

Returns a reference to the vector of values with the LYRICS key.

fn set_lyrics<T: Into<String>>(&mut self, lyrics: Vec<T>)

Sets the values for the LYRICS key.

fn remove_lyrics(&mut self)

Removes all values with the LYRICS key.